Audit Corrections Update
Zap Requests now update expiry timestamp to now.

This means:
For a Call or Put option is both parties agree to Zap, the Taker loses right to exercise the option in the temefram left from original expiry.

Writer can now delete the Option and pull out of a losing trade/scenario quicker.

Writer still faces a cancelation or deletion fee, the same that would have been charged to the winner of the trade as PayOff, but it measured in base tokens in this case.